+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/incremental/IndexBasedChecker.java
@@ -39,9 +39,9 @@ class IndexBasedChecker implements UpToDateChecker {
 		this.log = log;
 	}
 
-	static IndexBasedChecker create(MavenProject project, Iterable<Formatter> formatters, Log log) {
+	static IndexBasedChecker create(MavenProject project, Path indexFile, Iterable<Formatter> formatters, Log log) {
 		PluginFingerprint pluginFingerprint = PluginFingerprint.from(project, formatters);
-		FileIndexConfig indexConfig = new FileIndexConfig(project, pluginFingerprint);
+		FileIndexConfig indexConfig = new FileIndexConfig(project, indexFile, pluginFingerprint);
 		FileIndex fileIndex = FileIndex.read(indexConfig, log);
 		return new IndexBasedChecker(fileIndex, log);
 	}