+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/incremental/NoopChecker.java
@@ -24,8 +24,8 @@ class NoopChecker implements UpToDateChecker {
 
 	private NoopChecker() {}
 
-	static NoopChecker create(MavenProject project, Log log) {
-		FileIndexConfig indexConfig = new FileIndexConfig(project);
+	static NoopChecker create(MavenProject project, Path indexFile, Log log) {
+		FileIndexConfig indexConfig = new FileIndexConfig(project, indexFile);
 		FileIndex.delete(indexConfig, log);
 		return new NoopChecker();
 	}